requirements.txt
amqp==5.0.6
asgiref==3.2.10
attrs==21.2.0
billiard==3.6.4.0
cached-property==1.5.2
celery==5.0.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cligj==0.7.2
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
Django==3.1
django-celery-results==2.0.1
django-cors-headers==3.5.0
django-debug-toolbar==3.1.1
django-extensions==3.1.3
django-filter==2.4.0
djangorestframework==3.12.2
drf-yasg==1.20.0
gunicorn==20.0.4
idna==2.10
importlib-metadata==4.3.1
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
kiwisolver==1.3.1
kombu==5.1.0
MarkupSafe==2.0.1
matplotlib==3.3.4
munch==2.5.0
numpy==1.19.5
packaging==20.9
pandas==1.1.5
Pillow==8.1.0
prompt-toolkit==3.0.18
psycopg2-binary==2.8.3
Pygments==2.9.0
pyparsing==2.4.7
pyproj==3.0.1
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
redis==3.5.3
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
scipy==1.5.4
seaborn==0.11.1
Shapely==1.7.1
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.5
uWSGI==2.0.18
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.1
